+++ packages/terra-time-input/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 ## 4.45.1 - (March 1, 2022)
+  * Added
+    * Migrate docs to terra-framework-docs
 
 * Fixed
   * Supply isInvalid as well as aria-invalid to AccessibleInput